rename route, link admin dashboard to edit session
