+++ src/admin/EditDialogue.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const EditDialogue = () => {
+  return <div>EditDialogue</div>;
+};
+
+export default EditDialogue;
